fix(targetticket): possible SQL error
With a form configured to set location of a target ticket from a question, changing the question type to a text field causes attempt to set location from a string
instead if an ID. If the string contains a single quote, then the INSERT query is broken. This check the value is an integer. If not, lets ignore the value.

May need to lock questions when they are linked to a target.